1. add public_access bool field to instance table 2. skip route53 for non-public instances 3. python SQS publisher script injected via cloud-init to new spot 4. conditionally request public_ip when creating new spot 5. add example requests for web=server

+++ spot-manager.py
@@ -76,7 +76,7 @@ def delete_message(receipt_handle):
 # update_instance_record(body['name'], body['price'], body['ami'], body['type'], body['sg'], body['subnet'])
 #
 
-def update_instance_record(name, sir, price, ami, _type, sg, subnet, ns_record):
+def update_instance_record(name, sir, price, ami, _type, sg, subnet, public_access, ns_record):
     query = Instance.select().where(Instance.name == name)
     if query.exists():
         logger.info('updating record name: {}, with sir: {} params.'.format(name, sir))
@@ -88,10 +88,11 @@ def update_instance_record(name, sir, price, ami, _type, sg, subnet, ns_record):
         instance.sg = sg
         instance.subnet = subnet
         instance.ns_record = ns_record
+        instance.public_access = public_access
         instance.save()
     else:
         logger.info('adding new record: name: {} with sir: {} params.'.format(name, sir))
-        instance = Instance.create(name=name, ns_record=ns_record, sir=sir, price=price, ami=ami, type=_type, sg=sg,
+        instance = Instance.create(name=name, ns_record=ns_record, public_access=public_access, sir=sir, price=price, ami=ami, type=_type, sg=sg,
                                    subnet=subnet, state="spot-requested")
         instance.save()
 
@@ -109,8 +110,9 @@ def handle_incomming_message(body, receipt_handle, timestamp):
     # Must be False when running in prod. use True to clear
     # all messages in the SQS without handling ( 1 execution, then change back to False )
     #
-    # delete_all_queue_messages = True
+
     delete_all_queue_messages = False
+    #delete_all_queue_messages = True
 
     if delete_all_queue_messages:
         delete_message(receipt_handle)
@@ -130,22 +132,26 @@ def handle_incomming_message(body, receipt_handle, timestamp):
                 logger.info('updating record for: {} with public_ip: {}, private_ip: {}.'
                             .format(instance.name, instance.public_ip, instance.private_ip))
                 instance.save()
-                logger.info('updating route53 for: {} [ns_record: {} public_ip: {}].'
-                            .format(instance.name, instance.ns_record, instance.public_ip))
-                resp = change_ns_record(instance.name, instance.ns_record, instance.public_ip)
+                if instance.ns_record == 'NONE':
+                    logger.info('no DNS record for this spot, skipping route53')
+                else:
+                    logger.info('updating route53 for: {} [ns_record: {} public_ip: {}].'
+                                .format(instance.name, instance.ns_record, instance.public_ip))
+                    resp = change_ns_record(instance.name, instance.ns_record, instance.public_ip)
                 logger.info('deleting message.')
                 delete_message(receipt_handle)
             else:
                 logger.info('ignoring new spot with iid: {} (not managed by spot-manager).'.format(body['iid']))
 
         elif body['text'] == 'New spot requested from web-server':
-            logger.info('new spot request for {} : (price:{}, ami: {}, type: {}, sg: {}, subnet:{})'
-                        .format((body['name']), (body['price']), format(body['ami']), format(body['type']),format(body['sg']),format(body['subnet'])))
-            resp = request_new_spot(body['name'], body['price'], body['ami'], body['type'], body['sg'], body['subnet'])
+            logger.info('new spot request for {} : (price:{}, ami: {}, type: {}, sg: {}, subnet:{}, public_access: {})'
+                        .format((body['name']), (body['price']), format(body['ami']), format(body['type']),format(body['sg']),format(body['subnet']), format(body['public_access'])))
+
+            resp = request_new_spot(body['name'], body['price'], body['ami'], body['type'], body['sg'], body['subnet'], body['public_access'])
             # get the spot req id
             sir = resp['SpotInstanceRequests'][0]['SpotInstanceRequestId']
             update_instance_record(body['name'], sir, body['price'], body['ami'], body['type'], body['sg'],
-                                   body['subnet'], ns_record=body['ns_record'])
+                                   body['subnet'], body['public_access'], ns_record=body['ns_record'])
             logger.info('deleting message.')
             delete_message(receipt_handle)
         else:
@@ -203,8 +209,79 @@ def handle_incomming_message(body, receipt_handle, timestamp):
 
 
 # request_new_spot('0.05', 'ami-0fc8c8e37cd7db658', 'm5a.large', 'sg-056964e89bbf05266', 'subnet-03fb65f37827a8971'):
-def request_new_spot(name, spot_price, ami, _type, sg, subnet):
-    userdata_script = """#!/bin/bash
+def request_new_spot(name, spot_price, ami, _type, sg, subnet, public_access):
+    public_access = str.lower(public_access) == 'true'
+
+    sqs_config = """
+# export into the environment
+#
+
+export AWS_ACCESS_KEY_ID=AKIAJFWBUQHTFIVVMSHA
+export AWS_SECRET_ACCESS_KEY=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
+export AWS_DEFAULT_REGION=eu-west-1
+export QUEUE_URL=https://sqs.eu-west-1.amazonaws.com/390415077514/spot-test
+
+
+"""
+
+    publish_metadata_script = """#!/usr/bin/python3
+
+#
+# This script should run when instance first launched (metadata)
+# it will send a message to an SQS topic with the params (iid, private address, public address)
+#
+# remember to source env vars for aws access
+# for testing, you can override the message by running locally with args:
+#
+# ./publish-metadata.py 'this is a test message'
+#
+
+import boto3
+from ec2_metadata import ec2_metadata
+import json
+import sys
+import time
+import os
+
+# aws sqs parameters
+sqs = boto3.client('sqs',
+                   aws_access_key_id     = os.environ['AWS_ACCESS_KEY_ID'],
+                   aws_secret_access_key = os.environ['AWS_SECRET_ACCESS_KEY'],
+                   region_name           = os.environ['AWS_DEFAULT_REGION'])
+
+queue_url = os.environ['QUEUE_URL']
+version='1.10'
+
+# default message can be overiden by arg1
+if len(sys.argv) > 1:
+    text = sys.argv[1]
+else:
+    text = 'Lumos new spot instance launched'
+    
+# change None value if no public ip exists
+if ec2_metadata.public_ipv4 is None:
+    public_ip = 'NONE'
+else:
+    public_ip = ec2_metadata.public_ipv4
+
+
+msg = {
+    'text': text,
+    'iid':        ec2_metadata.instance_id,
+    'public_ip':  public_ip,
+    'private_ip': ec2_metadata.private_ipv4
+}
+
+resp = sqs.send_message(
+    QueueUrl=queue_url,
+    MessageBody=json.dumps(msg))
+
+#print(resp['MessageId'])
+print ('Message with own metadata has been sent to SQS.')
+    """
+
+
+    userdata_script = f"""#!/bin/bash
 
 #
 # DONT EXECUTE !
@@ -217,21 +294,47 @@ def request_new_spot(name, spot_price, ami, _type, sg, subnet):
 # 3. change DNS A records using route53 (mat be done externally also), if local need aws credentials
 # 4. change prompt (lms/root)
 
-LOG=/home/lms/user-data.output
-echo "$(date) - starting"              >> $LOG
-echo "executing userdata.sh"           >> $LOG
-echo "sourcing credentials"            >> $LOG
-source /home/lms/sqs-config
-echo "publish metadata to SQS"         >> $LOG
-python3 /home/lms/publish-metadata.py  >> $LOG
-echo "starting self redeployment"      >> $LOG
-/home/lms/self-deploy.sh               >> $LOG
-echo "$(date) - done."                 >> $LOG
-#echo "waiting 2m"                     >> $LOG
-#sleep 2m
-echo "waiting 20s before reboot"       >> $LOG
-sleep 20s
-reboot
+#LOG=/home/lms/user-data.output
+LOG=/home/ubuntu/user-data.output
+echo "$(date) - starting"                     >> $LOG
+echo "executing userdata.sh"                  >> $LOG
+
+
+echo "creating script to publish metadata"    >> $LOG
+cat << EOF > /home/ubuntu/publish-metadata.py
+{publish_metadata_script}
+EOF
+#echo "(publish_metadata_script)" > /home/ubuntu/publish-metadata.py
+chmod +x /home/ubuntu/publish-metadata.py
+
+echo "{sqs_config}"              > /home/ubuntu/sqs-config
+
+echo "installing python3"                     >> $LOG
+apt-get -yqq install python3-pip              >> $LOG
+pip3 install -q boto3 ec2_metadata            >> $LOG
+echo "python3 version: $(python3 -V)"         >> $LOG
+echo "pip3 version: $(pip3 -V)"               >> $LOG
+
+echo "sourcing credentials"                   >> $LOG
+#source /home/lms/sqs-config
+source /home/ubuntu/sqs-config
+echo "publish metadata to SQS"                >> $LOG
+#python3 /home/lms/publish-metadata.py        >> $LOG
+python3 /home/ubuntu/publish-metadata.py      >> $LOG
+#if [ -f "/home/lms/self-deploy.sh" ]; then
+if [ -f "/home/lms/self-deploy.sh" ]; then
+    echo "starting self redeployment"         >> $LOG
+    #/home/lms/self-deploy.sh                 >> $LOG
+    /home/ubuntu/self-deploy.sh               >> $LOG
+    echo "waiting 20s before reboot"          >> $LOG
+    sleep 20s
+    reboot
+else
+    echo "skipping self redeployment script." >> $LOG
+    echo "spot instance is ready."            >> $LOG
+fi
+echo "$(date) - done."                        >> $LOG
+
 """
     userdata_base64 = base64.b64encode(userdata_script.encode("ascii")).decode("ascii")
 
@@ -251,7 +354,9 @@ reboot
             'UserData': userdata_base64,
             'NetworkInterfaces': [
                 {
-                    'AssociatePublicIpAddress': True,
+                    'AssociatePublicIpAddress': public_access,
+                    #'AssociatePublicIpAddress': False,
+                    #'AssociatePublicIpAddress': True,
                     'DeviceIndex': 0,
                     'Groups': [sg],
                     'SubnetId': subnet,
